Split up server.rs (more)
* mod client: static files required for the client.
* mod rest: REST endpoints.
* mod html: plain HTML endpoints

also:
* Move Pagination into pagination mod.
